Nieuwe parser toegevoegd van David. TODO: else werkt niet (fout op de client)!

+++ svenserver/src/main/java/nl/virtualvikings/parser/Action.java
@@ -7,7 +7,7 @@ public class Action implements Statement {
 		this.action = action;
 	}
 	
-	public String result(byte[] stats) {
+	public String result(int[] stats) {
 		return action;
 	}
