+++ svenserver/src/main/java/nl/virtualvikings/parser/Assignment.java
@@ -3,13 +3,13 @@ package nl.virtualvikings.parser;
 public class Assignment implements Statement {
 	UserVariable left;
 	Expression right;
-	
+
 	public Assignment(UserVariable left, Expression right) {
 		this.left = left;
 		this.right = right;
 	}
 	
-	public String result(byte[] stats) {
+	public String result(int[] stats) {
 		left.setValue(new Constant(right.result(stats)));
 		return null;
 	}
