+++ svenserver/src/main/java/nl/virtualvikings/parser/Statement.java
@@ -1,6 +1,6 @@
 package nl.virtualvikings.parser;
 
 public interface Statement {
-	String result(byte[] stats);
+	String result(int[] stats);
 	String toString();
 }
